requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.1
filelock==3.14.0
fsspec==2024.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
openai==1.30.5
openai-whisper==20231117
pydantic==2.7.2
pydantic_core==2.18.3
regex==2024.5.15
requests==2.32.3
sniffio==1.3.1
sympy==1.12.1
tiktoken==0.7.0
torch==2.3.0
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.1
